split the main function and the libary, also modify the README
